+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
 
-# v2.0.0
-  - Change cli options and behaviour
+# v0.5.5
+
+- Parser v0.3.1
+
+  - Fix unresolved return type [05a1697137d8f1eed30b833cc86149646c6b344b](https://github.com/behemehal/Ellie-Language/commit/05a1697137d8f1eed30b833cc86149646c6b344b)
+  - Implement getter & setter [48ff89ad3e159f175bf7f4435caf4e47605719bc](https://github.com/behemehal/Ellie-Language/commit/48ff89ad3e159f175bf7f4435caf4e47605719bc)
+  - Fix faulty compiling sequence [931c9ffa930a19c42ad5186554ffafe18bdf2425](https://github.com/behemehal/Ellie-Language/commit/931c9ffa930a19c42ad5186554ffafe18bdf2425)
+
+- Tokenizer v0.2.2
+  - Fix cursor issues [3ea4d25726beae6fcecb00c477358cc6989055a3](https://github.com/behemehal/Ellie-Language/commit/3ea4d25726beae6fcecb00c477358cc6989055a3)
+  - [TECHNICAL] Save function's parameter name and type cursor positions separately [5ab7849dd09407baea6a39a47803a3c6e953314b](https://github.com/behemehal/Ellie-Language/commit/5ab7849dd09407baea6a39a47803a3c6e953314b)
+  - Fix the situation that ret cant be void. [97e877e5d08ca800e1c3c0394e09a158276a2d50](https://github.com/behemehal/Ellie-Language/commit/97e877e5d08ca800e1c3c0394e09a158276a2d50)
+  - [Major] Fix cursor issues with cli. [3ea4d25726beae6fcecb00c477358cc6989055a3](https://github.com/behemehal/Ellie-Language/commit/3ea4d25726beae6fcecb00c477358cc6989055a3)